+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * fixed bug on redirect from the old address of the "contact us" form;
 * fixed bug in getting the active users list. Now it shows the full name;
 * fixed bug on `StaticPage::paths()`. Now it returns only existing paths;
+* `LoginRecorder` is now a component;
 * added `AppTable::beforeSave()`. It checks if the `created` property is an
     instance of `Time`;
 * static pages uses cache. `StaticPage::all()` returns an array of `Entity`;